+++ test/composables/du-variant-persistence.test.ts
@@ -8,6 +8,7 @@ import { createAttaform } from '../../src/runtime/core/plugin'
 import { fingerprintZodSchema } from '../../src/runtime/adapters/zod-v4/fingerprint'
 import { hashStableString } from '../../src/runtime/core/hash'
 import type { UseFormReturnType } from '../../src/runtime/types/types-api'
+import { wait, waitUntil } from '../utils/form-harness'
 
 /**
  * Variant memory × persistence interaction. The contract under test:
@@ -173,29 +174,11 @@ function mount(persistKey: string): {
   return { app, api: handle.api as Api, setAddress, setNumber, setChannel }
 }
 
-async function wait(ms: number): Promise<void> {
-  await new Promise<void>((r) => setTimeout(r, ms))
-}
-
 async function drain(rounds = 8): Promise<void> {
   for (let i = 0; i < rounds; i++) {
     await Promise.resolve()
     await nextTick()
   }
-}
-
-async function waitUntil<T>(
-  predicate: () => T | null | undefined,
-  timeoutMs = 500,
-  intervalMs = 10
-): Promise<T> {
-  const start = Date.now()
-  while (Date.now() - start < timeoutMs) {
-    const v = predicate()
-    if (v !== null && v !== undefined) return v
-    await wait(intervalMs)
-  }
-  throw new Error('waitUntil: predicate never resolved')
 }
 
 describe('rememberVariants × persistence — refresh-survives contract', () => {
