+++ test/composables/use-form-v3-forwarding.test.ts
@@ -8,6 +8,7 @@ import { createAttaform } from '../../src/runtime/core/plugin'
 import { useForm } from '../../src/zod-v3'
 import { fingerprintZodSchema } from '../../src/runtime/adapters/zod-v3/fingerprint'
 import { hashStableString } from '../../src/runtime/core/hash'
+import { waitUntil } from '../utils/form-harness'
 
 /**
  * Regression pin: the zod v3 `useForm` wrapper used to hand-pick the
@@ -48,10 +49,6 @@ async function drain(rounds = 8): Promise<void> {
   }
 }
 
-async function wait(ms: number): Promise<void> {
-  await new Promise((resolve) => setTimeout(resolve, ms))
-}
-
 describe('v3 useForm forwards opt-in options to useAbstractForm', () => {
   const apps: App[] = []
   afterEach(() => {
@@ -76,7 +73,7 @@ describe('v3 useForm forwards opt-in options to useAbstractForm', () => {
     // reached useAbstractForm — populates fieldErrors within the
     // debounce window.
     api.setValue('email', 'nope')
-    await wait(60)
+    await waitUntil(() => (api.errors.email?.[0]?.message === 'bad email' ? true : null))
     await drain()
 
     expect(api.errors.email?.[0]?.message).toBe('bad email')
@@ -131,7 +128,7 @@ describe('v3 useForm forwards opt-in options to useAbstractForm', () => {
     if (input === undefined) throw new Error('email input not mounted')
     input.value = 'alice@example.com'
     input.dispatchEvent(new Event('input', { bubbles: true }))
-    await wait(50)
+    await waitUntil(() => (setItem.mock.calls.length > 0 ? true : null))
     await drain()
 
     expect(setItem).toHaveBeenCalled()
